Added in "list" command
Added in a new command, "list", to the ./securedb command-line tool, and an associated method AuthenticatedDBUser::get_record_names() to implement this command.

The simplicity of the added functionality belies a significant structural change that was required to implement it. In prior commits, all methods stored record names as hash values only, both as a means of identification and to prevent other users from seeing the record names. However, this meant that the actual owner cannot see their own record names either, which is not ideal.

In order to allow retrieval of record names, the Keys database was changed to contain four fields instead of three. The "record_name" field is now replaced with an encrypted version of the record name, created using crypto::encrypt() in conjunction with the user's master key. A field "record_identifier" was also added, which stores the hashed record name to be used as a secure identifier.

Details of how the C++ program files were modified to support this new structure and functionality are as follows:

dbmanager.h/dbmanager.cpp - modifications to AuthenticatedDBUser
* Added in get_record_names() method. Decrypts the names of the current user's records and returns them as a vector.
* Modified the public methods create_record(), delete_record() and private method get_record_key() to properly interface with the new DB structure.
* Added private method record_match(n) - finds owner record with name n.
* Minor edits to public methods such as retrieve_record() and edit_record() - mostly cleaning up variable names to make more sense with the new framework.

parsecmd.h/parsecmd.cpp
* Added in command type RECORDLIST, paired with user-typed command "list", and implemented parsing functionality for that command.

main.cpp
* Implemented the command "list" by calling AuthenticatedDBUser::get_record_names()

tests.cpp
* Added in unit tests related to get_record_names()
* Rewrote the pre-test database reset to properly comply with the new database structure.

./runtests was successfully run on the new edits prior to commit.